Fix polygroup problem
Problem was actually with polypaint parser. When poly paint was being read by the importer is was desynchronizing the parser. Polypaint parser now dynamically searches for the end of the polypaint data.

Disclaimer: Heavy use of AI code, parsing binary like this is out side of my skillset so I relied heavily on ai agents to make this fix.